rename to virtual, move block log transaction types, remove unnecessary types
