+++ examples/with-nextjs/ponder/ponder-env.d.ts
@@ -4,31 +4,24 @@
 // See https://ponder.sh/docs/guides/typescript for more information.
 
 declare module "@/generated" {
-  import type {
-    PonderApp,
-    PonderContext,
-    PonderEvent,
-    PonderEventNames,
-    PonderSchema,
-  } from "@ponder/core";
+  import type { Virtual } from "@ponder/core";
 
   type config = typeof import("./ponder.config.ts").default;
   type schema = typeof import("./ponder.schema.ts").default;
 
-  export const ponder: PonderApp<config, schema>;
-  export type EventNames = PonderEventNames<config>;
-  export type Event<name extends EventNames = EventNames> = PonderEvent<
+  export const ponder: Virtual.Registry<config, schema>;
+
+  export type EventNames = Virtual.EventNames<config>;
+  export type Event<name extends EventNames = EventNames> = Virtual.Event<
     config,
     name
   >;
-  export type Context<name extends EventNames = EventNames> = PonderContext<
+  export type Context<name extends EventNames = EventNames> = Virtual.Context<
     config,
     schema,
     name
   >;
-  export type IndexingFunctionArgs<name extends EventNames = EventNames> = {
-    event: Event<name>;
-    context: Context<name>;
-  };
-  export type Schema = PonderSchema<schema>;
+  export type IndexingFunctionArgs<name extends EventNames = EventNames> =
+    Virtual.IndexingFunctionArgs<config, schema, name>;
+  export type Schema = Virtual.Schema<schema>;
 }
