+++ packages/core/src/index.ts
@@ -1,15 +1,5 @@
 export { createConfig } from "@/config/config.js";
 export { createSchema } from "@/schema/schema.js";
-export type { Infer } from "@/schema/types.js";
-export type { Block } from "@/types/block.js";
-export type { Log } from "@/types/log.js";
-export type { DatabaseModel } from "@/types/model.js";
-export type {
-  PonderApp,
-  PonderContext,
-  PonderEvent,
-  PonderEventNames,
-  PonderSchema,
-} from "@/types/ponder.js";
-export type { Transaction } from "@/types/transaction.js";
 export { mergeAbis } from "@/utils/mergeAbis.js";
+export type { Block, Log, Transaction } from "@/types/eth.js";
+export type { Virtual } from "@/types/virtual.js";
