+++ packages/core/src/sync-store/store.ts
@@ -1,11 +1,8 @@
+import type { FactoryCriteria, LogFilterCriteria } from "@/config/sources.js";
+import type { Block, Log, Transaction } from "@/types/eth.js";
+import type { Checkpoint } from "@/utils/checkpoint.js";
 import type { Kysely, Migrator } from "kysely";
 import type { Address, Hex, RpcBlock, RpcLog, RpcTransaction } from "viem";
-
-import type { FactoryCriteria, LogFilterCriteria } from "@/config/sources.js";
-import type { Block } from "@/types/block.js";
-import type { Log } from "@/types/log.js";
-import type { Transaction } from "@/types/transaction.js";
-import type { Checkpoint } from "@/utils/checkpoint.js";
 
 export interface SyncStore {
   kind: "sqlite" | "postgres";
